move more classes into compose package
